requirements.txt
math 
numpy 
streamlit_option_menu
scipy.stats